+++ src/main/resources/release_notes.md
@@ -27,6 +27,10 @@
 
 * Portable version shows hint to install downloaded update manually (#2331)
 
+### Series
+
+* Fix error in table calculation if more than one team is replaced at one match day (#2453)
+
 ### Hall of fame
 
 * Added (#2296)
@@ -64,6 +68,7 @@
 
 * Fixed the missing translation for the home/away `Tactic` and `Details` column and displaying the abbreviation of the
   tactic now in the corresponding column.
+* Fixed a crash that occurred when not ticking `Only matches with SEs`.
 
 ### Option setting
 
@@ -92,11 +97,12 @@
 
 * Added the missing translation for the short text in the table for `Goals for the Team`.
 
-Reports by Contributors - October 19, 2025 - June 19, 2026
+Reports by Contributors - October 19, 2025 - September 11, 2026
 
 * Jösgall 1005
 * Ioannidouefty 459
-* wsbrenk 122
+* Lukasz 418
+* wsbrenk 130
 * Frankieorabona 73
 * Lidegand 72
 * \_KOHb\_ 54
@@ -107,4 +113,4 @@ Reports by Contributors - October 19, 2025 - June 19, 2026
 * Mara 3
 * Joeri Roels 1
 
-Total 1868
+Total 2294